experimentos/replicate-sdxl-test/requirements.txt
# Dependencias para Replicate + SDXL Test
# Instalación: pip install -r requirements.txt

# Replicate SDK (cliente oficial para Python)
replicate>=0.25.0

# Procesamiento de imágenes
Pillow>=10.0.0

# Requests para descargar imágenes
requests>=2.28.0

# ============================================================
# 📦 INSTALACIÓN RECOMENDADA:
# ============================================================
# python -m venv venv
# source venv/bin/activate  # En Windows: venv\Scripts\activate
# pip install -r requirements.txt
